bug(modal)
resolves bug with modal
