+++ client/components/protected/CreateGroup.jsx
@@ -1,0 +1,114 @@
+import React from 'react';
+import AppActions from '../../actions/AppActions';
+import * as Api from '../../api/AppApi';
+
+
+/**
+ * create creategroup components
+ *
+ * @class CreateGroup
+ *
+ * @extends {React.Component}
+ */
+class CreateGroup extends React.Component {
+  /**
+    * Creates an instance of CreateGroup
+    *
+    * @param {object} props
+    *
+    * @memberof CreateGroup
+  */
+  constructor(props) {
+    super(props);
+    this.state = {
+      groupName: '',
+      error: ''
+    };
+    this.onChange = this.onChange.bind(this);
+    this.onClick = this.onClick.bind(this);
+  }
+
+  /**
+  * @method onChange
+  *
+  * @description Monitors changes in the components and change the state
+  *
+  * @memberof CreateGroup
+  *
+  * @param {SyntheticEvent} event
+  *
+  * @returns {void}
+  */
+
+  onChange(event) {
+    this.setState({
+      groupName: event.target.value
+    });
+  }
+
+
+  /**
+   * @description creates a group if groupname is not empty
+   *
+   * @param {object} event
+   *
+   * @returns {void}
+   *
+   * @memberof CreateGroup
+  */
+  onClick(event) {
+    event.preventDefault();
+    const group = {
+      groupName: this.state.groupName,
+    };
+    if ((!this.state.groupName) || (!this.state.groupName.trim())) {
+      this.setState({
+        error: 'Please enter a valid group name'
+      });
+    } else {
+      AppActions.createGroup(group);
+      Api.getUserGroups();
+      this.setState({
+        groupName: '',
+        error: ''
+      });
+    }
+  }
+
+
+  /**
+   * @method render
+   *
+   * Render react component
+   *
+   * @returns {ReactElement} Create Group markup for the CreateGroup Components
+   *
+   * @memberof CreateGroup
+   */
+  render() {
+    return (
+      <div className="panel-body">
+        <div className="error"> {this.state.error} </div>
+        <form className="navbar-form" role="search">
+          <div className="form-group">
+            <input
+              type="text"
+              className="form-control"
+              placeholder="Create Group"
+              name="groupName"
+              onChange={this.onChange}
+              value={this.state.groupName}
+            />
+          </div>
+          <button
+            onClick={this.onClick}
+            type="submit"
+            id="submit"
+            className="btn btn-primary addMember"
+          > Submit </button>
+        </form>
+      </div>
+    );
+  }
+}
+export default CreateGroup;
